--- AlpsSDK/AlpsUser.swift
@@ -1,18 +1,0 @@
-//
-//  AlpsUser.swift
-//  Alps
-//
-//  Created by Rafal Kowalski on 21/12/2016
-//  Copyright © 2016 Alps. All rights reserved.
-//
-
-import Foundation
-import Alps
-
-open class AlpsUser {
-    public let manager: AlpsManager
-
-    public init(manager: AlpsManager) {
-        self.manager = manager
-    }
-}
